Changing filenames of logs

+++ xlnet_finetuning.py
@@ -104,12 +104,12 @@ def train(train_dataloader, val_dataloader, model, optimizer, scheduler, num_tra
 
 				if logging_step > 0 and global_step % logging_step == 0:
 					# Log metrics
-					eval_file_name = model_id + '_checkpoint_{}.json'.format(global_step/logging_step)
+					eval_file_name = model_id + '_checkpoint_{}.json'.format(int(global_step/logging_step))
 					results = evaluate(dataloader = val_dataloader, model = model, eval_file_name = eval_file_name, n_gpu=n_gpu, device=device)
 
 				if save_step > 0 and global_step % save_step == 0:
 					# Save model and optimizer checkpoints
-					checkpoint_save_path = os.path.join(model_save_path, 'model_checkpoint_{}'.format(global_step/save_step))
+					checkpoint_save_path = os.path.join(model_save_path, 'model_checkpoint_{}'.format(int(global_step/save_step)))
 					
 					if not os.path.exists(checkpoint_save_path):
 						os.makedirs(checkpoint_save_path)
